requirements.txt
matplotlib
numpy
onnxruntime
opencv-python==4.2.0.34
pyyaml
tqdm
